+++ application/android/app/src/main/java/com/group12/beabee/views/entities/QuestionFragmentEdit.java
@@ -10,7 +10,7 @@ import android.view.View;
 
 import com.group12.beabee.R;
 import com.group12.beabee.views.BaseInnerFragment;
-import com.group12.beabee.views.MainPage.PageMode;
+import com.group12.beabee.views.MainStructure.PageMode;
 
 /**
  * A simple {@link Fragment} subclass.
